+++ app/exchanges/binance/stream-client.js
@@ -70,6 +70,9 @@ _processMessage(message)
             case 'aggTrade':
                 this._processTrades(data);
                 return;
+            case 'kline':
+                this._processKline(data);
+                return;
         }
     }
     catch (e)
@@ -194,6 +197,55 @@ _processTrades(data)
         }]
     }
     this.emit('trades', evt);
+}
+
+/*
+  Process 'kline' data and emit 'kline' event
+
+  Example data
+
+  {
+    "e": "kline",     // Event type
+    "E": 123456789,   // Event time
+    "s": "BNBBTC",    // Symbol
+    "k": {
+      "t": 123400000, // Kline start time
+      "T": 123460000, // Kline close time
+      "s": "BNBBTC",  // Symbol
+      "i": "1m",      // Interval
+      "f": 100,       // First trade ID
+      "L": 200,       // Last trade ID
+      "o": "0.0010",  // Open price
+      "c": "0.0020",  // Close price
+      "h": "0.0025",  // High price
+      "l": "0.0015",  // Low price
+      "v": "1000",    // Base asset volume
+      "n": 100,       // Number of trades
+      "x": false,     // Is this kline closed?
+      "q": "1.0000",  // Quote asset volume
+      "V": "500",     // Taker buy base asset volume
+      "Q": "0.500",   // Taker buy quote asset volume
+      "B": "123456"   // Ignore
+    }
+  }
+
+ */
+_processKline(data)
+{
+    let pair = this._toCustomPair(data.s);
+    let evt = {
+        pair:pair,
+        interval:data.k.i,
+        data:{
+            timestamp:parseFloat(data.k.t / 1000.0),
+            open:parseFloat(data.k.o),
+            close:parseFloat(data.k.c),
+            high:parseFloat(data.k.h),
+            low:parseFloat(data.k.l),
+            volume:parseFloat(data.k.v)
+        }
+    }
+    this.emit('kline', evt);
 }
 
 }